services/optimizer/requirements.txt

# Amazon PPC Optimizer - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.28.0
pyyaml>=6.0
Flask>=3.0.0
gunicorn>=21.2.0

# Google Cloud integrations
google-cloud-bigquery>=3.17.0
google-cloud-secret-manager>=2.20.0

# Optional but recommended
python-dateutil>=2.8.2
pytz>=2023.3

# For enhanced logging and reporting
colorama>=0.4.6

# For data analysis (optional)
# pandas>=1.5.0
# numpy>=1.24.0
